+++ src/components/DropZone.tsx
@@ -1,6 +1,6 @@
-import React from "react";
 import { css, cx } from "emotion";
-import { useDrop, DragObjectWithType } from "react-dnd";
+import React from "react";
+import { DragObjectWithType, useDrop } from "react-dnd";
 
 const styles = {
   default: css`
