+++ tslint.json
@@ -1,0 +1,5 @@
+{
+  "rules": {
+    "ordered-imports": [true, { "grouped-imports": true }]
+  }
+}
